Basic Debug Panel implement
