+++ Assets/UI Toolkit/Stylesheets/Pause Page.uss.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8d634286dacda934cb669c06c1d9795f
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 12385, guid: 0000000000000000e000000000000000, type: 0}
+  disableValidation: 0
