+++ Assets/UI Toolkit/UI Pages/Pause Page.uxml.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 4c9ac4bb5739d5a4487ce86a444759f3
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
