+++ Assets/_Game/Scripts/Menus/Button Event.cs
@@ -1,4 +1,5 @@
 namespace fgames.Menus
 {
     public delegate void ButtonEvent();
+    public delegate void ToggleEvent(bool newValue);
 }
