+++ Assets/_Game/Scripts/Menus/Main Menu/Pages.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b27e90b5b34ce324fa02f4482776987b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
