+++ Assets/_Game/Scripts/Menus/Other.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0954af7df63d3944d923cb8ddce78150
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
