+++ Assets/_Game/Scripts/Menus/Pause/Pages.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 424d3cca43840784bb3e0af8b525d27a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
