+++ Assets/_Game/Scripts/Menus/Pause/Pages/Main Pause Page.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 11a1759baa4a2c34da277f3b8d830787
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
