+++ Assets/_Game/Scripts/Menus/Pause/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 13669666df884784ca0f37b65c1a073c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
