+++ Assets/_Game/Scripts/Menus/Utils.cs
@@ -11,5 +11,17 @@ namespace fgames.Menus
             newButton.text = text;
             return newButton;
         }
+
+        public static Toggle NewToggle(string text, bool value, Action<bool> onChange)
+        {
+            Toggle newToggle = new Toggle(text);
+            newToggle.value = value;
+            newToggle.RegisterCallback<ChangeEvent<bool>>((evt) => {
+                // UnityEngine.Debug.Log(evt.newValue);
+                // valueChanged?.Invoke(evt.newValue);
+                onChange(evt.newValue);
+            });
+            return newToggle;
+        }
     }
 }
