--- Assets/_Game/Scripts/Systems/Debug/Debug Page.cs
@@ -1,6 +1,0 @@
-using UnityEngine;
-
-public class DebugPage
-{
-
-}
