--- Assets/_Game/Scripts/Systems/Pause/Pause Page.cs
@@ -1,16 +1,0 @@
-using UnityEngine.UIElements;
-
-namespace fgames.Menus
-{
-    public class PauseScreenPage
-    {
-        private const string id_Body = "Body";
-        public string Title;
-        public VisualElement Body;
-
-        public PauseScreenPage(VisualElement root)
-        {
-            Body = root.Q<VisualElement>(id_Body);
-        }
-    }
-}
