--- Assets/_Game/Scripts/Systems/Pause/Pause UI.cs
@@ -1,16 +1,0 @@
-using UnityEngine;
-using UnityEngine.UIElements;
-
-namespace fgames.Menus
-{
-    public class PauseUI : MonoBehaviour
-    {
-        public UIDocument uiDocument;
-        private PauseScreenPage _page;
-
-        void Start()
-        {
-            _page = new PauseScreenPage(uiDocument.rootVisualElement);
-        }
-    }
-}
